+++ common/src/main/java/tocraft/craftedcore/data/PlayerDataSynchronizer.java
@@ -1,14 +1,13 @@
 package tocraft.craftedcore.data;
 
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.ListTag;
-import net.minecraft.nbt.StringTag;
-import net.minecraft.nbt.Tag;
+import net.minecraft.nbt.*;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import tocraft.craftedcore.CraftedCore;
 import tocraft.craftedcore.network.ModernNetworking;
 import tocraft.craftedcore.network.client.ClientNetworking;
+import tocraft.craftedcore.patched.CEntity;
 import tocraft.craftedcore.registration.PlayerDataRegistry;
 
 public class PlayerDataSynchronizer {
@@ -21,15 +20,33 @@ public class PlayerDataSynchronizer {
                 ListTag list = (ListTag) tag.get(PLAYER_DATA_SYNC);
                 if (list != null) {
                     for (Tag entry : list) {
-                        for (String key : ((CompoundTag) entry).getAllKeys()) {
-                            ClientNetworking.runOrQueue(context, player -> ((PlayerDataProvider) player).craftedcore$writeTag(key, ((CompoundTag) entry).get(key)));
+                        if (entry instanceof CompoundTag compoundTag) {
+                            for (String key : compoundTag.getAllKeys()) {
+                                ClientNetworking.runOrQueue(context, player -> {
+                                    PlayerDataProvider playerDataProvider;
+                                    if (tag.hasUUID("uuid")) {
+                                        playerDataProvider = (PlayerDataProvider) player.getCommandSenderWorld().getPlayerByUUID(tag.getUUID("uuid"));
+                                    } else {
+                                        playerDataProvider = (PlayerDataProvider) player;
+                                    }
+                                    if (playerDataProvider != null) {
+                                        playerDataProvider.craftedcore$writeTag(key, ((CompoundTag) entry).get(key));
+                                    }
+                                });
+                            }
+                        } else if (entry instanceof StringTag stringTag) {
+                            ClientNetworking.runOrQueue(context, player -> {
+                                PlayerDataProvider playerDataProvider;
+                                if (tag.hasUUID("uuid")) {
+                                    playerDataProvider = (PlayerDataProvider) player.getCommandSenderWorld().getPlayerByUUID(tag.getUUID("uuid"));
+                                } else {
+                                    playerDataProvider = (PlayerDataProvider) player;
+                                }
+                                if (playerDataProvider != null) {
+                                    playerDataProvider.craftedcore$writeTag(stringTag.getAsString(), null);
+                                }
+                            });
                         }
-                    }
-                }
-                ListTag deletedTags = (ListTag) tag.get(PLAYER_DATA_SYNC + "_DEL");
-                if (deletedTags != null) {
-                    for (Tag deletedTag : deletedTags) {
-                        ClientNetworking.runOrQueue(context, player -> ((PlayerDataProvider) player).craftedcore$writeTag(deletedTag.getAsString(), null));
                     }
                 }
             }
@@ -40,28 +57,56 @@ public class PlayerDataSynchronizer {
      * Synchronize data from the server to the client
      */
     public static void sync(ServerPlayer player) {
+        syncToSelf(player);
+        syncToAll(player);
+    }
+
+    private static void syncToSelf(ServerPlayer player) {
         CompoundTag tag = new CompoundTag();
         ListTag list = new ListTag();
-        ListTag deletedTags = new ListTag();
 
         PlayerDataProvider playerData = ((PlayerDataProvider) player);
 
         for (String key : ((PlayerDataProvider) player).craftedcore$keySet()) {
             // ignore key if it shouldn't be synchronized to the client
-            if (!PlayerDataRegistry.shouldSyncKey(key))
+            if (!PlayerDataRegistry.shouldSyncTagToSelf(key)) {
                 return;
+            }
+
+            CompoundTag entry = new CompoundTag();
+            Tag value = playerData.craftedcore$readTag(key);
+            entry.put(key, value);
+            list.add(entry);
+        }
+        tag.put(PLAYER_DATA_SYNC, list);
+
+        ModernNetworking.sendToPlayer(player, PLAYER_DATA_SYNC_ID, tag);
+    }
+
+    private static void syncToAll(ServerPlayer player) {
+        CompoundTag tag = new CompoundTag();
+        ListTag list = new ListTag();
+
+        tag.putUUID("uuid", player.getUUID());
+        PlayerDataProvider playerData = ((PlayerDataProvider) player);
+
+        for (String key : ((PlayerDataProvider) player).craftedcore$keySet()) {
+            // ignore key if it shouldn't be synchronized to the client
+            if (!PlayerDataRegistry.shouldSyncTagToAll(key)) {
+                return;
+            }
 
             CompoundTag entry = new CompoundTag();
             Tag value = playerData.craftedcore$readTag(key);
             if (value != null) {
                 entry.put(key, value);
-                list.add(entry);
             } else {
-                deletedTags.add(StringTag.valueOf(key));
+                list.add(StringTag.valueOf(key));
             }
+            list.add(entry);
         }
         tag.put(PLAYER_DATA_SYNC, list);
-        tag.put(PLAYER_DATA_SYNC + "_DEL", deletedTags);
-        ModernNetworking.sendToPlayer(player, PLAYER_DATA_SYNC_ID, tag);
+        //noinspection resource
+        ModernNetworking.sendToPlayers(((ServerLevel) CEntity.level(player)).players(), PLAYER_DATA_SYNC_ID, tag);
     }
 }